chore(ring-health): deduplicate code betwwen cortex and loki
